requirements.txt
Pillow
requests
matplotlib